requirements.txt
streamlit
numpy<2.0.0
torch
torchvision
pandas
matplotlib
pillow
plotly
opencv-python-headless
requests
PyYAML
tqdm
seaborn
scipy
psutil
ultralytics